Move into one script
